Add typings for set_options.
